requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
scikit-learn>=1.3.0
nltk>=3.8.0
textblob>=0.17.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
lxml>=4.9.0
fake-useragent>=1.4.0
waitress>=2.1.2
